add provisioning database user and don't recompile php every time

+++ provision/provision.sh
@@ -26,24 +26,29 @@ yum install -y httpd mysql-server httpd-devel patch libpng-devel apr-devel libxm
 service mysqld start
 /usr/bin/mysqladmin -u root password 'blank'
 
-mkdir prefork
-git clone https://github.com/Automattic/prefork.git ./prefork
+if [ -f /usr/local/bin/php ]
+then
+	echo "PHP already installed, skipping compiling"
+else
+	mkdir prefork
+	git clone https://github.com/Automattic/prefork.git ./prefork
 
-mkdir /etc/php.d
-wget http://us.php.net/get/php-5.4.17.tar.gz/from/us3.php.net/mirror
-tar -zxvf php-*.tar.gz
-cd php-5.4.17
-patch  -p 1 -i ../prefork/php_cli.c.diff
-./configure --with-config-file-path=/etc --with-config-file-scan-dir=/etc/php.d --with-apxs2 --with-libdir=lib64 --enable-mbstring --with-openssl --with-soap --with-mcrypt --with-gd --with-libevent --with-memcached --with-pear --with-mysql --with-mysqli --with-xml --with-zlib
-make clean
-make
-make install
-/home/vagrant/php-5.4.17/libtool --finish /home/vagrant/php-5.4.17/libs
-cp /home/vagrant/php-5.4.17/php.ini-recommended /etc/php.ini
-printf "\n" | /usr/local/bin/pecl install libevent-beta
-echo "extension=libevent.so" >> /etc/php.ini
-cp /vagrant/php.conf /etc/httpd/conf.d/php.conf
-service httpd restart
+	mkdir /etc/php.d
+	wget http://us.php.net/get/php-5.4.17.tar.gz/from/us3.php.net/mirror
+	tar -zxvf php-*.tar.gz
+	cd php-5.4.17
+	patch  -p 1 -i ../prefork/php_cli.c.diff
+	./configure --with-config-file-path=/etc --with-config-file-scan-dir=/etc/php.d --with-apxs2 --with-libdir=lib64 --enable-mbstring --with-openssl --with-soap --with-mcrypt --with-gd --with-libevent --with-memcached --with-pear --with-mysql --with-mysqli --with-xml --with-zlib
+	make clean
+	make
+	make install
+	/home/vagrant/php-5.4.17/libtool --finish /home/vagrant/php-5.4.17/libs
+	cp /home/vagrant/php-5.4.17/php.ini-production /etc/php.ini
+	printf "\n" | /usr/local/bin/pecl install libevent-beta
+	echo "extension=libevent.so" >> /etc/php.ini
+	cp /vagrant/php.conf /etc/httpd/conf.d/php.conf
+	service httpd restart
+fi
 
 # Get WordPress
 wpclone=/usr/local/src/wordpress
@@ -73,6 +78,10 @@ then
 		printf "   ...Failed to copy WordPress ${site}.dev files\n"
 	else
 		printf "   ...${site_name}.dev files copied\n"
+		#Create WordPress database
+		mysql -u root -pblank -e 'create database prefork'
+		mysql -u root -pblank -e "create user 'wp'@'localhost' identified by 'wp'"
+		mysql -u root -pblank -e "grant all on prefork.* to 'wp'@'localhost'"
 	fi
 else
 	printf " * Skip setting up files at ${location}, already setup\n"
